Monstro, PianoRoll improvements

+++ plugins/monstro/Monstro.cpp
@@ -1593,14 +1593,14 @@ void MonstroInstrument::updateSamplerate()
 void MonstroInstrument::updateSlope1()
 {
 	const float slope = m_env1Slope.value();
-	m_slope1 = powf( 10.0f, static_cast<double>( slope * -1.0f ) ); 
+	m_slope1 = powf( 10.0f, slope * -1.0f ); 
 }
 
 
 void MonstroInstrument::updateSlope2()
 {
 	const float slope = m_env2Slope.value();
-	m_slope2 = powf( 10.0f, static_cast<double>( slope * -1.0f ) ); 
+	m_slope2 = powf( 10.0f, slope * -1.0f ); 
 }
